+++ src/components/Navbar/TopWidgets/TopActions.vue
@@ -1,18 +1,5 @@
 <script setup lang="ts">
-import ConfirmDeleteDialog from '@/components/Dialogs/ConfirmDeleteDialog.vue'
-import { useDashboardStore } from '@/stores'
-import { ref } from 'vue'
-
-defineEmits(['resumeTorrents', 'pauseTorrents', 'openSearchEngine', 'openrssArticles', 'openLogs', 'openSettings'])
-
-const dashboardStore = useDashboardStore()
-
-const deleteTorrentDialogVisible = ref(false)
-const deleteHashes = ref<string[]>()
-function onDelete() {
-  deleteTorrentDialogVisible.value = true
-  deleteHashes.value = [...dashboardStore.selectedTorrents]
-}
+defineEmits(['resumeTorrents', 'pauseTorrents', 'deleteTorrents', 'openSearchEngine', 'openrssArticles', 'openLogs', 'openSettings'])
 </script>
 
 <template>
@@ -30,13 +17,7 @@ function onDelete() {
 
   <v-tooltip :text="$t('topbar.actions.delete')" location="bottom">
     <template v-slot:activator="{ props }">
-      <v-btn icon v-bind="props" @click="onDelete">
-        <v-icon>mdi-delete</v-icon>
-        <ConfirmDeleteDialog v-if="deleteTorrentDialogVisible"
-                             v-model="deleteTorrentDialogVisible"
-                             :hashes="deleteHashes"
-                             disable-activator />
-      </v-btn>
+      <v-btn icon="mdi-delete" v-bind="props" @click="$emit('deleteTorrents')" />
     </template>
   </v-tooltip>
